Git - expose random name generation for branches (#298938)
